requirements.txt
Django==4.2.20
djangorestframework==3.15.2
djangorestframework-simplejwt
django-cors-headers
django-filter
drf-yasg
Pillow
requests
gunicorn
sqlparse
tzdata
coreapi
